Refactoring copy button code block
