+++ src/components/MdxComponents.tsx
@@ -8,6 +8,7 @@ import { ProductCard, ProductCards } from './mdx/ProductCards';
 import { QuickLink, QuickLinks } from './mdx/QuickLinks';
 import { Tab, Tabs } from './mdx/Tabs';
 import ResourceEstimator from './resource-estimator/ResourceEstimator';
+import { PreCodeBlock } from './PreCodeBlock';
 
 const MdxComponents = (version?: string) => {
 	return {
@@ -50,7 +51,8 @@ const MdxComponents = (version?: string) => {
 		},
 		video: (props: any) => {
 			return <video className="rounded-xl" {...props} />;
-		}
+		},
+		pre: (props: any) => <PreCodeBlock {...props} />
 	};
 };
