+++ src/components/PreCodeBlock.tsx
@@ -1,0 +1,27 @@
+import { CopyButton } from '@/components/CopyButton'
+import React, { DetailedHTMLProps, HTMLAttributes, Children } from 'react'
+
+interface PreProps extends DetailedHTMLProps<HTMLAttributes<HTMLPreElement>, HTMLPreElement> {
+  raw?: string
+}
+
+export function PreCodeBlock ({ children, ...props }: PreProps) {
+  const propsObj = { ...props }
+  const propsValues = Object.values(propsObj)
+  const [, , dataLanguage, dataTheme] = propsValues
+  const lang = dataLanguage || "shell"
+  
+  let codeContent = ''
+  Children.forEach(children, child => {
+    if (typeof child === 'string') codeContent += child
+  })
+
+  return (
+    <pre data-language={lang} data-theme={dataTheme} className={"py-4"}>
+      <div className='relative'>
+        <CopyButton text={props?.raw || ''}  />
+        <div className={"p-2"}>{children}</div>
+      </div>
+    </pre>
+  )
+}
